refactor(query): simplify Oracle plan lifecycle
